consumer/pyproject.toml
[project]
name = "consumer"
version = "0.1.0"
dependencies = [
    "confluent-kafka",
    "python-dotenv",
    "sqlalchemy"
]
